Put documentation build behind checkbox
Fixes #597

+++ new/postinstall.sh
@@ -36,12 +36,14 @@ do
 done
 
 # JSDoc build (ignore errors on earlier MediaWiki versions)
-cd $PATCHDEMO/wikis/$NAME/w
-npm ci || true
-npm run doc || true
-# remove node_modules afterwards, they take up infinite disk space and won't be needed any more
-rm -rf $PATCHDEMO/wikis/$NAME/w/node_modules || true
-cd $PATCHDEMO
+if [ "${BUILD_DOCS}" = "1" ]; then
+	cd $PATCHDEMO/wikis/$NAME/w
+	npm ci || true
+	npm run doc || true
+	# remove node_modules afterwards, they take up infinite disk space and won't be needed any more
+	rm -rf $PATCHDEMO/wikis/$NAME/w/node_modules || true
+	cd $PATCHDEMO
+fi
 
 # OOUI build
 if [ -d $PATCHDEMO/wikis/$NAME/w/build/ooui ]; then
